requirements.txt
pandas>=2.2
numpy>=1.26
scipy>=1.11         
scikit-learn>=1.4    
scikit-surprise>=1.1 
joblib>=1.4          
matplotlib>=3.9      